mnis_extra

A wrapper for mnis_additional functions. It combines the various options of mnis_additional into one dataframe, and the default is similar to mnis_full_biog. Variable descriptions are taken from the mnis website: <http://data.parliament.uk/membersdataplatform/memberquery.aspx>

mnis_extra(ID, mnis_id = TRUE, ref_dods = FALSE, addresses = TRUE,
  basic_details = TRUE, biography_entries = TRUE, committees = TRUE,
  constituencies = TRUE, elections_contested = TRUE, experiences = TRUE,
  government_posts = TRUE, honours = TRUE, house_memberships = TRUE,
  interests = TRUE, known_as = TRUE, maiden_speeches = TRUE,
  opposition_posts = TRUE, other_parliaments = TRUE,
  parliamentary_posts = TRUE, parties = TRUE, preferred_names = TRUE,
  staff = TRUE, statuses = TRUE, tidy = TRUE)

Arguments

ID
The ID number of the member. Defaults to NULL.

mnis_id
Request based on the default membership ID scheme.

ref_dods
Request based on the DODS membership ID scheme. Defaults to FALSE.

addresses
Member address information (e.g. website, twitter, consituency address etc...).

basic_details
Basic information about the Member (e.g. given name, HoL membership type, oaths etc...)

biography_entries
Member biographical information (e.g. countries of interest, policy expertise etc...)

committees
Committees a Member sits or has sat on as well details on committee chairing.

constituencies
constituencies a Member has represented.

elections_contested
Elections a Member has contested but not won.

experiences
Non-parliamentary experience of a Member.

government_posts
Government posts a Member currently holds.

honours
Honours (e.g. MBE, OBE etc...) held by a Member.

house_memberships
House membership list of a Member.

interests
Registered interests (financial) of a Member.

known_as
Details of names a Member has chosen to be known as instead of their full title (House of Lords members only).

maiden_speeches
Maiden speech dates for a Member.

opposition_posts
Opposition posts a Member has held.

other_parliaments
Other Parliaments that a Member has held a membership of.

parliamentary_posts
Parliamentary posts a Member has held.

parties
Party affiliations of a Member.

preferred_names
Full set of data about a Members' name (e.g. surname, forename, Honorary prefixes, full details of HoL title and rank etc...).

staff
The staff employed by a Member.

statuses
Status history (e.g. suspensions and disqualifications) for a Member.

tidy
Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to snake_case. Defaults to TRUE.

Value

A tibble with data on a given MP.

See also

mnis_full_biog mnis_basic_details mnis_additional

Examples

not_run({

  x <- mnis_extra(172)

})