Peers on date

Requests data on all peers who were members of the House of Lords on the date specified, (if only date1 is included as a parameter), or on or between the two dates if both date1 and date2 are specified. Either date1 or date2 can be the latter of the two dates.

mnis_peers_on_date(
  date1 = Sys.Date(),
  date2 = NULL,
  tidy = TRUE,
  tidy_style = "snake_case"
)

Arguments

date1

The date to return the list of peers from. Defaults to current system date. Accepts character values in 'YYYY-MM-DD' format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with as.Date().

date2

An optional query parameter. Accepts character values in 'YYYY-MM-DD' format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with as.Date(). If not NULL, the function returns a list of all peers in the House of Lords between date1 and date2. Defaults to NULL.

tidy

If TRUE, fixes the variable names in the tibble to remove non-alphanumeric characters and superfluous text, and convert to a consistent style. Defaults to TRUE.

tidy_style

The style to convert variable names to, if tidy=TRUE. Accepts one of "snake_case", "camelCase" and "period.case". Defaults to "snake_case".

Value

A tibble with information on all peers who were members of the House of Lords on the date specified, (if only date1 is included as a parameter), or on or between the two dates if both date1and date2 are specified.

See also

Examples

if (FALSE) {
x <- mnis_peers_on_date(date1 = "2017-01-01")
}