Political and Geographical Interests

Returns a tibble with members with the specified interest or interests.

mnis_political_interests(
  interest,
  house = NULL,
  current = TRUE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

Arguments

interest
A single interest in the form of a string, or a character vector of interests. If a vector of multiple interests, returns all members who list all of those interests.

house
The house the member belongs to. Accepts 'commons', 'lords' and NULL. This variable is not case sensitive. If NULL, returns members from both houses. Defaults to NULL.

current
Logical. If TRUE, returns all current and former members of both houses. Defaults to TRUE.

tidy
If TRUE, fixes the variable names in the tibble to remove non-alphanumeric characters and superfluous text, and convert to a consistent style. Defaults to TRUE.

tidy_style
The style to convert variable names to, if tidy=TRUE. Accepts one of "snake_case", "camelCase" and "period.case". Defaults to "snake_case".

Value

A data frame of members with given interest(s).

Examples

if (FALSE) {
x <- mnis_political_interests(c("Africa", "Jamaica"))

y <- mnis_political_interests(c("Africa", "Europe"))

z <- mnis_political_interests(c("Africa", "Education"))
}