Additional member information

A series of basic function for the API lookup. Each of these functions accepts a member's ID and returns information; if no ID is given basic information on all members of both houses is returned. All functions return basic details about the member (name, date of birth, gender, constituency, party, IDs, current status, etc.), as well as any available additional information requested by the specific function.

mnis_basic_details(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_biography_entries(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_committees(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_addresses(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_constituencies(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_elections_contested(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_experiences(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_government_posts(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_honours(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_house_memberships(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_statuses(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_staff(ID = NULL, ref_dods = FALSE, tidy = TRUE, tidy_style = "snake_case")

mnis_interests(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_known_as(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_maiden_speeches(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_opposition_posts(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_other_parliaments(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_parliamentary_posts(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_parties(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

mnis_preferred_names(
  ID = NULL,
  ref_dods = FALSE,
  tidy = TRUE,
  tidy_style = "snake_case"
)

Arguments

ID
The member ID. If NULL, function calls mnis_all_members() and returns basic information on all members of both houses. Defaults to NULL.

ref_dods
If TRUE, Request based on the DODS membership ID scheme. If FALSE, requests data based on the default membership ID scheme. Defaults to FALSE.

tidy
If TRUE, fixes the variable names in the tibble to remove non-alphanumeric characters and superfluous text, and convert to a consistent style. Defaults to TRUE.

tidy_style
The style to convert variable names to, if tidy=TRUE. Accepts one of "snake_case", "camelCase" and "period.case". Defaults to "snake_case".

Value

A list with the data corresponding to the particular function called.

mnis_additional functions

mnis_basic_details Basic biographical details on a given Member

mnis_biography_entries Member biographical information (e.g. countries of interest, policy expertise etc...)

mnis_committees Committees a Member sits or has sat on as well details on committee chairing

mnis_addresses Member address information (e.g. website, twitter, constituency address etc...)

mnis_constituencies Constituencies a Member has represented

mnis_elections_contested Elections a Member has contested but not won

mnis_experiences Non-parliamentary experience of a Member

mnis_government_posts Government posts a Member has held

mnis_honours Honours (e.g. MBE, OBE etc...) held by a Member

mnis_house_memberships House membership list of a Member

mnis_statuses Status history (e.g. suspensions and disqualifications) for a Member

mnis_staff The staff employed by a Member

mnis_interests Registered (financial) interests of a Member

mnis_known_as Details of names a Member has chosen to be known as instead of their full title, only applicable to members of the House of Lords

mnis_maiden_speeches Maiden speech dates for a Member

mnis_opposition_posts Opposition posts a Member has held

mnis_other_parliaments Other Parliaments that a Member has held a membership of (e.g. members of the Welsh Assembly)

mnis_parliamentary_posts Parliamentary posts a Member has held

mnis_parties Party affiliations of a Member

mnis_preferred_names Full set of data about a Members' name (e.g. surname, forename, Honorary prefixes, full details of House of Lords title and rank if applicable, etc...)

See also

Examples

if (FALSE) {
x <- mnis_basic_details(172)
}