mnis_additional

A series of basic function for the API lookup.

mnis_additional()

mnis_basic_details(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_biography_entries(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_committees(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_addresses(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_constituencies(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_elections_contested(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_experiences(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_government_posts(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_honours(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_house_memberships(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_statuses(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_staff(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_interests(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_known_as(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_maiden_speeches(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_opposition_posts(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_other_parliaments(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_parliamentary_posts(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_parties(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_preferred_names(ID = NULL, ref_dods = FALSE, tidy = TRUE)

Arguments

ID
The member ID value. If empty, function calls mnis_all_members and returns basic information on all members of both houses.

ref_dods
Request based on the DODS membership ID scheme. Defaults to FALSE, where it requests data based on the default membership ID scheme.

tidy
Fix the variable names in the tibble to remove non-alphanumeric characters and superfluous text. Defaults to TRUE.

Value

A tibble with the data corresponding to the particular function called.

See also

mnis_full_biog mnis_extra

Examples

not_run({

  x <- mnis_basic_details(172)

})
not_run({

  x <- mnis_additional()

})