A wrapper for mnis_additional functions.

It combines the various options of mnis_additional into one dataframe, and the default is similar to mnis_full_biog. Variable descriptions are taken from the mnis website: <http://data.parliament.uk/membersdataplatform/memberquery.aspx>.

mnis_extra(ID, ref_dods = FALSE, addresses = TRUE,
  biography_entries = TRUE, committees = TRUE, constituencies = TRUE,
  elections_contested = TRUE, experiences = TRUE, government_posts = TRUE,
  honours = TRUE, house_memberships = TRUE, interests = TRUE,
  known_as = TRUE, maiden_speeches = TRUE, opposition_posts = TRUE,
  other_parliaments = TRUE, parliamentary_posts = TRUE, parties = TRUE,
  preferred_names = TRUE, staff = TRUE, statuses = TRUE, tidy = TRUE,
  tidy_style = "snake_case")

Arguments

ID
The ID number of the member. If NULL, does not return any data. Defaults to NULL.

ref_dods
Request based on the DODS membership ID scheme. Defaults to FALSE. If FALSE, requests data based on the default membership ID scheme. Defaults to TRUE.

addresses
Member address information (e.g. website, twitter, consituency address etc...). If TRUE, address details are included in the tibble. Defaults to TRUE.

biography_entries
Member biographical information (e.g. countries of interest, policy expertise etc...) If TRUE, biographical details are included in the tibble. Defaults to TRUE.

committees
Committees a Member sits or has sat on as well details on committee chairing. If TRUE, committee details are included in the tibble. Defaults to TRUE.

constituencies
Constituencies a Member has represented. If TRUE, constituency details are included in the tibble. Defaults to TRUE.

elections_contested
Elections a Member has contested but not won. If TRUE, details of unsuccessful election contests are included in the tibble. Defaults to TRUE.

experiences
Non-parliamentary experience of a Member. If TRUE, extra-parliamentary experience details are included in the tibble. Defaults to TRUE.

government_posts
Government posts a member has held. If TRUE, government posts details are included in the tibble. Defaults to TRUE.

honours
Honours (e.g. MBE, OBE etc...) held by a Member. If TRUE, honours details are included in the tibble. Defaults to TRUE.

house_memberships
House membership list of a Member. If TRUE, house membership details are included in the tibble. Defaults to TRUE.

interests
Registered interests (financial) of a Member. If TRUE, interest details are included in the tibble. Defaults to TRUE.

known_as
Details of names a Member has chosen to be known as instead of their full title (House of Lords members only). If TRUE, known as details are included in the tibble. Defaults to TRUE.

maiden_speeches
Maiden speech dates for a Member. If TRUE, maiden speech details are included in the tibble. Defaults to TRUE.

opposition_posts
Opposition posts a Member has held. If TRUE, opposition post details are included in the tibble. Defaults to TRUE.

other_parliaments
Other Parliaments that a Member has held a membership of. If TRUE, details of other parliaments are included in the tibble. Defaults to TRUE.

parliamentary_posts
Parliamentary posts a Member has held. If TRUE, parliamentary posts details are included in the tibble. Defaults to TRUE.

parties
Party affiliations of a Member. If TRUE, party details are included in the tibble. Defaults to TRUE.

preferred_names
Full set of data about a Members' name (e.g. surname, forename, Honorary prefixes, full details of HoL title and rank etc...). If TRUE, preferred names details are included in the tibble. Defaults to TRUE.

staff
The staff employed by a Member. If TRUE, staff details are included in the tibble. Defaults to TRUE.

statuses
Status history (e.g. suspensions and disqualifications) for a Member. If TRUE, status details are included in the tibble. Defaults to TRUE.

tidy
If TRUE, fixes the variable names in the tibble to remove non-alphanumeric characters and superfluous text, and convert to a consistent style. Defaults to TRUE.

tidy_style
The style to convert variable names to, if tidy=TRUE. Accepts one of 'snake_case', 'camelCase' and 'period.case'. Defaults to 'snake_case'.

Value

A tibble with the requested data on a given MP.

See also

mnis_full_biog mnis_basic_details mnis_additional

Examples

# NOT RUN {
x <- mnis_extra(172)
# }